added category as a seperate app
