+++ config/urls.py
@@ -2,7 +2,6 @@
 from django.contrib import admin
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-
 from v1.shop.urls import router as shop_router
 
 urlpatterns = [
@@ -12,7 +11,6 @@ urlpatterns = [
 ]
 
 router = DefaultRouter(trailing_slash=False)
-
 router.registry.extend(shop_router.registry)
 urlpatterns += router.urls
 #urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
