+++ v1/category/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from .models import Category
+
+admin.site.register(Category)
