+++ v1/category/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CategoryConfig(AppConfig):
+    name = 'category'
