+++ v1/category/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from .models import Category, Item
+
+class CategorySerializer(serializers.ModelSerializer):
+  class Meta:
+    model = Category
+    fields = '__all__'
