+++ v1/category/urls.py
@@ -1,0 +1,2 @@
+from django.urls import path
+from . import views
