+++ v1/item/admin.py
@@ -1,8 +1,7 @@
 from django.contrib import admin
 
-from .models import Category, Item, Cart, CartItem
+from .models import Item, Cart, CartItem
 
-admin.site.register(Category)
 admin.site.register(Item)
 admin.site.register(Cart)
 admin.site.register(CartItem)
